Set up descriptors for USB HID.
Set up for boot protocol keyboard and mouse.

The keyboard descriptor worked out to be exactly 64 bytes, which seems
to have triggered a remaining bug in chopstx ZLP handling.

The HID interface requests are just stubbed in for now to get the
devices to not error out when connected to Windows/Linux.