Keyboard now works.
Added header with USB HID usage codes, and mapping of Sun 'make codes'
to USB HID usage codes (from suniversal, with the addition of 0x0f which
is the 'blank' key on my keyboard, where escape is on 'normal'
keyboards).

Added code to populate keyboard hid report and transmit it to the host.