Add TODO comment for 'phantom state'.
According to the HID spec, when there are more keys pressed than space
in the report, all entries in the array must be the Usage(ErrorRollOver)
code, and once the number of keys down is reduced to where they would
fit the array is filled with the keys that are actually down.  This
would require the firmware to track ALL key states somewhere, and seems
like a lot of effort and memory for something pretty uncommon.